requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
duckduckgo_search==7.2.1
e2b==1.0.5
e2b-code-interpreter==1.0.3
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.58.4
lxml==5.3.0
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.1
openai==1.59.8
orjson==3.10.14
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
primp==0.10.1
propcache==0.2.1
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smolagents==1.4.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.1
yarl==1.18.3
yfinance==0.2.51
zipp==3.21.0
